+++ src/gui/web-components/util/IconWebComponent.ts
@@ -26,29 +26,7 @@ export class IconWebComponent extends HTMLElement {
             return;
         }
 
-        const iconMap: { [key: string]: string } = {
-            //airplane: "✈️",
-            alarm: "⏰",
-            //check: "✓",
-            //clipboard: "📋", // metar
-            //"clipboard-check": "🛫", // flightplan
-            //clock: "🕑",
-            //"cloud-haze": "🌁",
-            //"cloud-sun": "⛅",
-            //clouds: "☁️",
-            compass: "🧭",
-            //folder: "📁",
-            //"fuel-pump": "⛽",
-            //gear: "⚙️",
-            //"hourglass-split": "⏳",
-            //info: "ℹ️",
-            //"thermometer-half": "🌡️",
-            //"three-dots-vertical": "⋮",
-            //wind: "🧭",
-            //"x": "✗",
-        };
-
-        this.innerHTML = iconMap[newValue] ?? this.getHtmlForSvg(newValue);
+        this.innerHTML = this.getHtmlForSvg(newValue);
     }
 
     private getHtmlForSvg(icon: string) {